requirements.txt
# Core dependencies
gradio==4.16.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
pillow>=9.5.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Deep learning utilities
ninja>=1.11.0
scipy>=1.10.0
lpips>=0.1.4
dlib>=19.24.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Data handling
pandas>=2.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
